feat: add trophy
