+++ content/0.index.md
@@ -6,7 +6,7 @@ image: /socials/index.jpg
 
 ::home-hero
 ---
-sectionClass: mt-24
+sectionClass: mt-24 relative overflow-hidden
 topButtonText: S'inscrire à l'Afterwork Battle
 topButtonLink: afterworkBattleRegistrationLink
 primaryButtonText: Découvrir l'aventure
@@ -16,12 +16,14 @@ secondaryButtonLink: concoursRegistrationLink
 ---
 
 #title
-Le Classement<br />des Associations
+[Le Classement<br />des Associations]{.relative.z-[1]}
 
 #subtitle
-L'aventure humaine de référence<br />
-qui [rassemble]{.font-semibold}, [promeut]{.font-semibold} et [valorise]{.font-semibold}<br />
-la vie associative étudiante.
+[L'aventure humaine de référence<br />]{.relative.z-[1]}
+[qui [rassemble]{.font-semibold}, [promeut]{.font-semibold} et [valorise]{.font-semibold}<br />]{.relative.z-[1]}
+[la vie associative étudiante.]{.relative.z-[1]}
+#extra
+  :home-trophy{.absolute.right-0}
 ::
 
 ::home-sponsors
