--- public/assets/home/trophy.mtl
@@ -1,14 +1,0 @@
-# WaveFront *.mtl file (generated by Cinema 4D)
-
-newmtl Or
-Ka 1 1 1
-Kd 1 1 1
-illum 7
-
-newmtl Mat
-Ka 1 1 1
-Kd 0.80000001192093 0.80000001192093 0.80000001192093
-Ks 1 1 1
-Ns 50
-illum 7
-
